Phase D: the Hermes return channel works, and album/artist notes use it
Server side is in place and verified end to end: four routes (music-notes,
music-lyrics, music-vibe, music-health) plus `~/.hermes/bin/nd`, the wrapper that
reads Navidrome's database read-only, calls the API with credentials from
~/.hermes/.env, and writes the results file atomically. A signed health request
came back through the mailbox in about twenty seconds.

Writing `nd` is what made the prompts honest, and two of its queries were wrong
until the database said so. `media_file.genre` is empty for all 25,784 rows --
Navidrome 0.63 keeps tags as JSON in `media_file.tags` -- so genre filtering had
to go through `json_each`. And the naive duplicate query put NIN's seventeen
`[silence]` tracks at the top; the useful query is titles that exist in *both*
music folders, which is 4,290 groups and the only case where "which copy do I
delete" has an answer. Undecodable now uses the same suffix list the player
transcodes on, so the report and the app agree: 11 .ogg files, exactly the figure
the plan predicted.

On the client, asking is not the same as requesting, so it is a separate service.
An acquisition is answered by music appearing in the library -- that is what
MusicRequestService watches -- while these are answered by text that only exists
in the mailbox. Same signing, same secret, same client, different definition of
done. 404 stays "still working": the timeout is per feature, because a note takes
a minute and a playlist can take twenty.

Notes fill a hole that is visible today: this server's getArtistInfo2 returns no
biography and its image URLs 404, so an artist page is a grid and nothing else.
Cached with explicit load/store rather than the fetch-first helper, since a note
costs an agent run and never changes -- and failures are never cached, or one bad
minute would mean that record can never have a note again.

+++ Tests/MusicLogicTests/TripPlannerTests.swift
@@ -11,7 +11,7 @@ struct TripPlannerTests {
         var song = Song(id: id, title: title ?? "Track \(id)")
         song.artist = artist
         song.duration = minutes * 60
-        song.size = minutes * 60 * 1_000_000
+        song.size = Int64(minutes * 60 * 1_000_000)
         return song
     }
